+++ DataGenerator/Data/Zones/ZoneInfoBase.cs
@@ -1381,6 +1381,26 @@ namespace DataGenerator.Data
                         layout.GenSteps.Add(PR_GRID_GEN, step);
                     }
 
+                    //special room gen
+                    {
+                        RoomGenLoadMap<MapGenContext> loadRoom = new RoomGenLoadMap<MapGenContext>();
+                        loadRoom.MapID = "test_room";
+                        loadRoom.RoomTerrain = new Tile(0);
+                        int width = 8;
+                        int height = 8;
+                        loadRoom.Borders = new Dictionary<Dir4, bool[]>
+                        {
+                            [Dir4.Down] = new bool[width],
+                            [Dir4.Up] = new bool[width],
+                            [Dir4.Left] = new bool[height],
+                            [Dir4.Right] = new bool[height],
+                        };
+                        SetGridSpecialRoomStep<MapGenContext> specialStep = new SetGridSpecialRoomStep<MapGenContext>();
+                        specialStep.Rooms = new PresetPicker<RoomGen<MapGenContext>>(loadRoom);
+                        specialStep.RoomComponents.Set(new ImmutableRoom());
+                        layout.GenSteps.Add(PR_GRID_GEN_EXTRA, specialStep);
+                    }
+
                     AddDrawGridSteps(layout);
 
 
@@ -1400,7 +1420,6 @@ namespace DataGenerator.Data
                     structure.Floors.Add(layout);
                 }
                 #endregion
-
 
                 #region CHASM CAVE
                 {