+++ CHANGELOG.md
@@ -5,4 +5,5 @@
 * Fixed autotiling issue with StableTiles
 * Add developer option to make ground maps wrap around.  Dungeon map wrapping is WIP.
 * Added documentation for move VFX and map editor
+* Maps can be loaded as custom rooms into a floor generator.  Use RoomGenLoadMap
 * Death is now processed in Universal, with PreDeath and SetDeath events.  Mods that modify Universal.bin will need to update OnDeath with these events.